minor renames, tweaks
