+++ ffp-backend/FPA_FOD_PLUS/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from .models import Data
+
+# Register your models here.
+admin.site.register(Data)
