+++ ffp-backend/FPA_FOD_PLUS/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class FFPConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'FPA_FOD_PLUS'
