+++ ffp-backend/FPA_FOD_PLUS/tests.py
@@ -1,0 +1,3 @@
+from django.test import TestCase
+
+# Create your tests here.
