+++ ffp-backend/FPA_FOD_PLUS/urls.py
@@ -1,0 +1,11 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path('', views.index, name='index'),
+    
+    path('results/', views.results, name='results'),
+    
+    path('administrator/', views.administrator, name='administrator'),
+]
