--- ffp-backend/data2019/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-
-from .models import Data
-
-# Register your models here.
-admin.site.register(Data)
