--- ffp-backend/data2019/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class Data2019Config(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'data2019'
