--- ffp-backend/data2019/urls.py
@@ -1,11 +1,0 @@
-from django.urls import path
-
-from . import views
-
-urlpatterns = [
-    path('', views.index, name='index'),
-    
-    path('results/', views.results, name='results'),
-    
-    path('administrator/', views.administrator, name='administrator'),
-]
