--- ffp-backend/data2019/views.py
@@ -1,20 +1,0 @@
-from django.http import HttpResponse
-
-from .models import Data
-
-def index(request):
-    
-    return HttpResponse("Hello, world. You're at the data2019 index page.")
-    
-def results(request):
-    query_results = Data.objects.all()
-    response = ""
-    for row in query_results:
-        response += "FOD_ID: " + str(row.FOD_ID)
-        response += " FPA_ID: " + str(row.FPA_ID)
-        response += " NAME: " + str(row.NAME) + "\n"
-    
-    return HttpResponse(response)
-    
-def administrator(request):
-    return HttpResponse("Hello you are looking at the administrator page.")
