+++ ffp-backend/ffp-backend/urls.py
@@ -17,6 +17,6 @@ from django.contrib import admin
 from django.urls import include, path
 
 urlpatterns = [
-    path('data2019/', include('data2019.urls')),
+    path('data/', include('FPA_FOD_PLUS.urls')),
     path('admin/', admin.site.urls),
 ]
