--- ffp-backend/polls/admin.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-
-from .models import Question
-
-# Register your models here.
-
-admin.site.register(Question)
